requirements.txt
# Core scientific computing libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# HTTP client (used for internal calc engine testing only)
requests>=2.31.0

# gRPC dependencies
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0

# Web framework for API endpoints
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# JSON handling
orjson>=3.9.0

# Production dependencies
gunicorn>=21.2.0

# Development dependencies (optional, uncomment for local development)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
